Added Files
Objective:  Write code to display the number of days, given the month 

There is no time requirement, but Phase 1 should take an hour or less and Phase 2 should take another 30 minutes to an hour.  This is an optional challenge, but a couple of hours of your time could earn up to 5 points on the next exam.

The problem is broken into phases. Phase 1 earns up to 3 points. Phase 2 earns an additional 2 points. Points will be added to Exam 2. 



Phase 1 (3 points):  Implement the following program in MIPS:
* In the main program:
* get month from the user
* if the user enters 0, exit the program
* if the month is not 1-12, print an error message and repeat
* print the month and number of days as shown in the sample run
* Write a function to determine the number of days by implementing a switch statement that is logically the same as shown in the box to the right, but in MIPS. 



Sample Phase 1 run (your wording can vary):







Phase 2 (2 additional points): Instead of printing the number of the month, give the user the option to either print the number or print the 3-character abbreviation. Ask the user only once at the beginning of the program. Keep the user in a loop until they enter 1 or 2. Phase 2 sample run (Welcome message is optional):


	

Upload your .asm file to eLearning by Friday noon
This is an individual task, programs will be checked for similarity across sections.